+++ confiture-rest-api/prisma/migrations/20221212230747_add_audit_creation_date_field/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Audit" ADD COLUMN     "creationDate" TIMESTAMP(3);