+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Tous les changements notables de Ara sont documentés ici avec leur date, leur c
   - Contenus non soumis à l’obligation d’accessibilité
 - Ajout d'une explication sur le calcul du taux de conformité sur le rapport ([#232](https://github.com/DISIC/Ara/pull/232))
 - Cache les sections "Contenus non accessibles" vides dans le rapport
+- Ajoute un message d'information sur la page de rapport d'un audit en cours ([#236](https://github.com/DISIC/Ara/pull/236)).
 
 ### Corrections